tests: adds deploy locally to e2e tests
