+++ .github/workflows/tests.yml
@@ -13,7 +13,7 @@ jobs:
     strategy:
       matrix:
         # Disabled windows because it's too slow
-        os: [ubuntu-latest, macos-latest]
+        os: [macos-latest]
         node: ["18"]  
     name: Run tests on ${{ matrix.os }}
     runs-on: ${{ matrix.os }}
