+++ tests/fixture-projects/lock/package.json
@@ -1,11 +1,28 @@
 {
     "name": "lock",
     "dependencies": {
-        "@nomicfoundation/hardhat-ethers": "^3.0.8"
-    },
-    "pnpm": {
-        "overrides": {
-            "@parity/revive>solc": "0.8.29"
-        }
-  }
+        "@nomicfoundation/hardhat-chai-matchers": "^2.0.8",
+        "@nomicfoundation/hardhat-ethers": "^3.0.8",
+        "@nomicfoundation/hardhat-ignition": "^0.15.10",
+        "@nomicfoundation/hardhat-ignition-ethers": "^0.15.10",
+        "@nomicfoundation/hardhat-network-helpers": "^1.0.12",
+        "@nomicfoundation/hardhat-toolbox": "^5.0.0",
+        "@nomicfoundation/hardhat-verify": "^2.0.13",
+        "@nomicfoundation/ignition-core": "^0.15.10",
+        "@openzeppelin/hardhat-upgrades": "^3.9.0",
+        "@parity/resolc": "^0.1.0-dev.16",
+        "@typechain/ethers-v6": "^0.5.1",
+        "@typechain/hardhat": "^9.1.0",
+        "@types/chai": "^4.2.0",
+        "@types/mocha": ">=9.1.0",
+        "@types/node": ">=18.0.0",
+        "chai": "^4.2.0",
+        "ethers": "^6.13.5",
+        "hardhat": "2.22.0",
+        "hardhat-gas-reporter": "^1.0.8",
+        "solidity-coverage": "^0.8.14",
+        "ts-node": ">=8.0.0",
+        "typechain": "^8.3.2",
+        "typescript": ">=4.5.0"
+    }
 }
