+++ tests/helpers.sh
@@ -58,3 +58,16 @@ assert_directory_not_empty() {
     exit 1
   fi
 }
+
+check_log_value() {
+    local output="$1"
+    local expected_value="$2"
+    
+    echo "$output"
+
+    if echo "$output" | grep -q "$expected_value"; then
+      :
+    else
+        exit 1
+    fi
+}
